TransactionalMasterRequestListener: refectored for better extansibility
- created TransactionalMasterRequestConditionProviderInterface that is used to determinate whether a transaction should be initialized
