+++ packages/framework/src/Component/HttpFoundation/TransactionalMasterRequestListener.php
@@ -5,26 +5,23 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Component\HttpFoundation;
 
 use Doctrine\ORM\EntityManagerInterface;
-use GraphQL\Error\SyntaxError;
-use GraphQL\Language\AST\OperationDefinitionNode;
-use GraphQL\Language\Parser;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
+use Traversable;
 
 class TransactionalMasterRequestListener
 {
-    protected const GRAPHQL_ENDPOINT_ROUTE = 'overblog_graphql_endpoint';
-    protected const QUERY_TYPE = 'query';
-
     protected bool $inTransaction = false;
 
     protected static bool $isManuallyRollbacked = false;
 
     /**
+     * @param \Traversable<int, \Shopsys\FrameworkBundle\Component\HttpFoundation\TransactionalMasterRequestConditionProviderInterface> $transactionalMasterRequestConditionProviders
      * @param \Doctrine\ORM\EntityManagerInterface $em
      */
     public function __construct(
+        protected readonly Traversable $transactionalMasterRequestConditionProviders,
         protected readonly EntityManagerInterface $em,
     ) {
     }
@@ -39,7 +36,7 @@ class TransactionalMasterRequestListener
      */
     public function onKernelRequest(RequestEvent $event): void
     {
-        if ($event->isMainRequest() && !$this->inTransaction && !$this->isRequestGraphQlQuery($event)) {
+        if (!$this->inTransaction && $this->shouldBeginTransaction($event)) {
             $this->em->beginTransaction();
             $this->inTransaction = true;
         }
@@ -50,7 +47,7 @@ class TransactionalMasterRequestListener
      */
     public function onKernelResponse(ResponseEvent $event): void
     {
-        if ($event->isMainRequest() && $this->inTransaction && !static::$isManuallyRollbacked) {
+        if ($this->inTransaction && !static::$isManuallyRollbacked) {
             $this->em->commit();
             $this->inTransaction = false;
         }
@@ -61,48 +58,24 @@ class TransactionalMasterRequestListener
      */
     public function onKernelException(ExceptionEvent $event): void
     {
-        if ($event->isMainRequest() && $this->inTransaction) {
+        if ($this->inTransaction) {
             $this->em->rollback();
             $this->inTransaction = false;
         }
     }
 
     /**
-     * @param \Symfony\Component\HttpKernel\Event\RequestEvent $requestEvent
+     * @param \Symfony\Component\HttpKernel\Event\RequestEvent $event
      * @return bool
      */
-    protected function isRequestGraphQlQuery(RequestEvent $requestEvent): bool
+    protected function shouldBeginTransaction(RequestEvent $event): bool
     {
-        if ($requestEvent->getRequest()->attributes->get('_route') !== static::GRAPHQL_ENDPOINT_ROUTE) {
-            return false;
-        }
-
-        $requestContent = $requestEvent->getRequest()->getContent();
-
-        if ($requestContent === null || $requestContent === '') {
-            return false;
-        }
-
-        $source = json_decode($requestContent, true);
-
-        if (!array_key_exists(static::QUERY_TYPE, $source)) {
-            return false;
-        }
-
-        $queryString = $source[static::QUERY_TYPE];
-
-        try {
-            $parsed = Parser::parse($queryString);
-
-            foreach ($parsed->definitions as $definition) {
-                if ($definition instanceof OperationDefinitionNode) {
-                    return $definition->operation === static::QUERY_TYPE;
-                }
+        foreach ($this->transactionalMasterRequestConditionProviders as $transactionalMasterRequestConditionProvider) {
+            if (!$transactionalMasterRequestConditionProvider->shouldBeginTransaction($event)) {
+                return false;
             }
-        } catch (SyntaxError) {
-            return false;
         }
 
-        return false;
+        return true;
     }
 }
